+++ src/stam_annotator/loaders/opa_loader.py
