+++ src/stam_annotator/loaders/opf_loader.py
